Make the source optional
This commit adds the ability to allow the module to be used without
deploying the source.

A new parameter was created: deploy_source

If set to false, the module will assume the source of the installation
is already present and extracted at the specified "tmp_dir" path.

If set to true (default), the module will attempt to fetch and extract
the source.