Handle updating software.. sorta
This commit adds the semi-ability to update IBM software by killing
processes that are running out of the target directory.  IBM
Installation Manager requires that all related services be stopped
before making software modifications to a target directory.
Unfortunately, they don't seem to provide a great way of doing it other
than going through the instance directory and knowing what to stop.  We
search the process table for items that match the target directory and
send a 'kill' to it.  This ain't pretty.

Also corrects the reporting of version.  Since we can't have this thing
upgrade a version if a newer is specified (because too many human tasks
need to be done before we can automate that - like
downloading/extracting the software), we instead reach satisfaction if
the current or greater version is present.