Fixes idempotency problem
Fixes an idempotency problem when using ibm_pkg to apply a fixpack onto the
same target directory. IBM Installation Manager will overwrite the version
in the installed.xml file with the fixpack version, causing the prefetch
to not pick up the older version. Changing this to retrieve the install
data from installRegistry.xml fixes it.