(MODULES-4738) add param :jdk_package_name
Currently, when using imcl install, you can only specify one package name. For installing WebSphere 9.x, you must also pass the jdk package name in the same line. This inserts the name of the jdk package into the command where it needs to be alongside the other package name if it is not nil.